Add abstraction for sequence numbers
This class represents a sequence number, which only can take on
values between [0, 102400]. This invariant is enforced by overloading
operaotr '+' to perform modulo arithmetic. Additionally, the standard
comparison operators are overloaded so that these can be treated closer
to normally integers. The comparison methods rely solely on the difference
method, which is the only thing complex here.

The difference between 2 sequence numbers encodes which of the 2 are
greater. Standard subtraction can be used when the sequence numbers
are "close" together, and otherwise the numbers need to be "adjusted"
in order to be compared properly.