Implement client retransmission and proper congestion control
The implementation maintains a list of Packet objects to be sent.
They are removed only until an ACK arrives (todo: only remove
them from the list if the seqno matches the ack number). The
retransmission timer resets every time a new ACK arrives, and
upon expiration, the client sends the first Packet in the list.