implement fin packets in the server
This also handles receiving the fin packets out of order, by buffering
fin packets like normal data packets. In addition, the server knows
whether to expect an ACK or not now, by keeping track of pending SYN
and FIN segments. This could later be used to implement retransmission
in the server.